[launcher] Use deferred-eval strings instead of arrays
Values are stored as un-expanded variable expansions, i.e. '${1#-J}' and
expanded when needed. The original positional parameters are kept as
long as possible so they can be referenced by the "arrays". Currently
there's no way to store an expansion with whitespace, but if necessary
the whitespace could be stored in its own variable.